feat(llc): Basic LLC exit valuation calculator
Python3 script that takes in csv of investors and outputs the multiplier each
investor will get on their investment for a series of exit valuations. Takes into account:
1. Founder's stock
2. PI (only one per investor currently)
3. PI Valuation
4. Monetary Investment
5. Time investment (based off of yearly salary lost)
6. Time multiplier (how much time does this person put in relative to baseline)

This script loads a file named 'data.csv' and looks for the above columns. It outputs
'total.csv' with the first row being the valutations and subsequent rows each of the
investors. First set is multiplier on investment, second set is net gain (or loss).